requirements.txt
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
psycopg2==2.6.2
connexion==1.0.109